tried changing double to float for speed; didn't work

+++ bpf.c
@@ -13,30 +13,30 @@
 #include "uniform.h"
 #include "gaussian-erfinv.h"
 
-typedef struct { double x, y; } ccoord;
-typedef struct { double r, t; } acoord;
+typedef struct { float x, y; } ccoord;
+typedef struct { float r, t; } acoord;
 typedef struct { ccoord posn; acoord vel; } state;
-typedef struct { state state; double weight; } particle_info;
+typedef struct { state state; float weight; } particle_info;
 
-typedef particle_info *resample(double);
+typedef particle_info *resample(float);
 
 static resample resample_naive, resample_optimal;
 
-static const double nsecs = 100;
-static const double dt = 0.1;
+static const float nsecs = 100;
+static const float dt = 0.1;
 
 int nparticles = 100;
 int sort = 0;
 static resample *resampler = resample_naive;
 
-static double avar = M_PI / 16;
-static double pvar = 0.1;
+static float avar = M_PI / 16;
+static float pvar = 0.1;
 
-static void update_state(state *p, double dt) {
-    double r0 = p->vel.r + gaussian(pvar);
-    double t0 = p->vel.t + gaussian(avar);
-    double x0 = p->posn.x + r0 * cos(t0) * dt;
-    double y0 = p->posn.y - r0 * sin(t0) * dt;
+static void update_state(state *p, float dt) {
+    float r0 = p->vel.r + gaussian(pvar);
+    float t0 = p->vel.t + gaussian(avar);
+    float x0 = p->posn.x + r0 * cos(t0) * dt;
+    float y0 = p->posn.y - r0 * sin(t0) * dt;
 
     p->vel.r = r0;
     p->vel.t = t0;
@@ -71,7 +71,7 @@ ccoord gps_measure(void) {
     return result;
 }
 
-acoord imu_measure(double dt) {
+acoord imu_measure(float dt) {
     acoord result = vehicle.vel;
     result.r += gaussian(pvar / dt);
     result.t += gaussian(avar / dt);
@@ -79,27 +79,27 @@ acoord imu_measure(double dt) {
 }
 
 /* XXX is this right? */
-double gprob(double delta, double sd) {
-    return exp(-(delta * delta * sd));
+float gprob(float delta, float sd) {
+    return expf(-(delta * delta * sd));
 }
 
-double gps_prob(state *s, ccoord *gps) {
-    double px = gprob(s->posn.x - gps->x, pvar);
-    double py = gprob(s->posn.y - gps->y, pvar);
+float gps_prob(state *s, ccoord *gps) {
+    float px = gprob(s->posn.x - gps->x, pvar);
+    float py = gprob(s->posn.y - gps->y, pvar);
     /* XXX is this right? */
     return px * py;
 }
 
-double imu_prob(state *s, acoord *imu) {
-    double pr = gprob(s->vel.r - imu->r, pvar);
-    double pt = gprob(s->vel.t - imu->t, avar);
+float imu_prob(state *s, acoord *imu) {
+    float pr = gprob(s->vel.r - imu->r, pvar);
+    float pt = gprob(s->vel.t - imu->t, avar);
     /* XXX is this right? */
     return pr * pt;
 }
 
-double sum_weights(void) {
+float sum_weights(void) {
     int i;
-    double t = particle[0].weight;
+    float t = particle[0].weight;
     for (i = 1; i < nparticles; i++)
 	t += particle[i].weight;
     return t;
@@ -113,10 +113,10 @@ int argmax_weight(void) {
     return i;
 }
 
-particle_info *weighted_sample(double scale) {
+particle_info *weighted_sample(float scale) {
     int i;
-    double w = uniform() * scale;
-    double t = 0;
+    float w = uniform() * scale;
+    float t = 0;
     for (i = 0; i < nparticles; i++) {
 	t += particle[i].weight;
 	if (t >= w)
@@ -126,7 +126,7 @@ particle_info *weighted_sample(double scale) {
     abort();
 }
 
-particle_info *resample_naive(double scale) {
+particle_info *resample_naive(float scale) {
     particle_info *newp = particle_states[!which_particle];
     int i;
     for (i = 0; i < nparticles; i++)
@@ -134,15 +134,15 @@ particle_info *resample_naive(double scale) {
     return newp;
 }
 
-double nform(int n) {
+float nform(int n) {
     return 1.0 - pow(uniform(), 1.0 / (n + 1));
 }
 
-particle_info *resample_optimal(double scale) {
+particle_info *resample_optimal(float scale) {
     particle_info *newp = particle_states[!which_particle];
-    double u0 = nform(nparticles - 1) * scale;
+    float u0 = nform(nparticles - 1) * scale;
     int i, j = 0;
-    double t = 0;
+    float t = 0;
     for (i = 0; i < nparticles; i++) {
         while (t + particle[j].weight < u0 && j < nparticles)
 	    t += particle[j++].weight;
@@ -153,7 +153,7 @@ particle_info *resample_optimal(double scale) {
     return newp;
 }
 
-inline static int sgn(double x) {
+inline static int sgn(float x) {
     if (x < 0)
 	return -1;
     if (x > 0)
@@ -167,7 +167,7 @@ static int cmp_weight(const void *w1, const void *w2) {
     return -sgn(p1->weight - p2->weight);
 }
 
-state bpf_step(ccoord *gps, acoord *imu, double dt) {
+state bpf_step(ccoord *gps, acoord *imu, float dt) {
     int i;
     particle_info *newp;
     /* update particles */
@@ -175,12 +175,12 @@ state bpf_step(ccoord *gps, acoord *imu, double dt) {
 	update_state(&particle[i].state, dt);
     /* do probabilistic weighting */
     for (i = 0; i < nparticles; i++) {
-	double gp = gps_prob(&particle[i].state, gps);
-	double ip = imu_prob(&particle[i].state, imu);
+	float gp = gps_prob(&particle[i].state, gps);
+	float ip = imu_prob(&particle[i].state, imu);
         particle[i].weight = gp * ip;
     }
     /* get normalizing constant */
-    double tweight = sum_weights();
+    float tweight = sum_weights();
     /* resample */
     if (sort)
 	qsort(particle, nparticles, sizeof(particle[0]), cmp_weight);
@@ -194,7 +194,7 @@ state bpf_step(ccoord *gps, acoord *imu, double dt) {
 }
 
 void run(void) {
-    double t;
+    float t;
     init_particles();
     for(t = 0; t <= nsecs; t += dt) {
 	update_state(&vehicle, dt);
